[Unit Tests] Get tests to pass while test machine is offline.
