+++ Kiosk/App/Networking/ArtsyAPI.swift
@@ -331,11 +331,10 @@ func endpointResolver() -> ((endpoint: Endpoint<ArtsyAPI>) -> (NSURLRequest)) {
 class ArtsyProvider<T where T: MoyaTarget>: ReactiveCocoaMoyaProvider<T> {
     typealias OnlineSignalClosure = () -> RACSignal
 
-    // Closure that returns a signal which completes once the app is online.
-    let onlineSignal: OnlineSignalClosure
+    let onlineSignal: RACSignal
 
-    init(endpointClosure: MoyaEndpointsClosure = MoyaProvider.DefaultEndpointMapping, endpointResolver: MoyaEndpointResolution = MoyaProvider.DefaultEndpointResolution, stubBehavior: MoyaStubbedBehavior = MoyaProvider.NoStubbingBehavior, networkActivityClosure: Moya.NetworkActivityClosure? = nil, onlineSignal: OnlineSignalClosure = connectedToInternetOrStubbingSignal) {
-        self.onlineSignal = onlineSignal
+    init(endpointClosure: MoyaEndpointsClosure = MoyaProvider.DefaultEndpointMapping, endpointResolver: MoyaEndpointResolution = MoyaProvider.DefaultEndpointResolution, stubBehavior: MoyaStubbedBehavior = MoyaProvider.NoStubbingBehavior, networkActivityClosure: Moya.NetworkActivityClosure? = nil, onlineSignal: RACSignal = connectedToInternetOrStubbingSignal()) {
+        self.onlineSignal = onlineSignal.replayLast()
         super.init(endpointClosure: endpointClosure, endpointResolver: endpointResolver, stubBehavior: stubBehavior, networkActivityClosure: networkActivityClosure)
     }
 }
@@ -366,7 +365,7 @@ struct Provider {
     }
     
     static func StubbingProvider() -> ArtsyProvider<ArtsyAPI> {
-        return ArtsyProvider(endpointClosure: endpointsClosure, endpointResolver: endpointResolver(), stubBehavior: MoyaProvider.ImmediateStubbingBehaviour, onlineSignal: { RACSignal.empty() })
+        return ArtsyProvider(endpointClosure: endpointsClosure, endpointResolver: endpointResolver(), stubBehavior: MoyaProvider.ImmediateStubbingBehaviour, onlineSignal: RACSignal.`return`(true))
     }
 
     private struct SharedProvider {
