+++ Kiosk/App/Networking/XAppAuthentication.swift
@@ -28,7 +28,7 @@ private func XAppTokenRequest(defaults: NSUserDefaults) -> RACSignal {
 /// Request to fetch a given target. Ensures that valid XApp tokens exist before making request
 func XAppRequest(token: ArtsyAPI, provider: ArtsyProvider<ArtsyAPI> = Provider.sharedProvider, defaults: NSUserDefaults = NSUserDefaults.standardUserDefaults()) -> RACSignal {
 
-    return provider.onlineSignal().then {
+    return provider.onlineSignal.ignore(false).take(1).then {
         // First perform XAppTokenRequest(). When it completes, then the signal returned from the closure will be subscribed to.
         XAppTokenRequest(defaults).then {
             return provider.request(token)
